[fix] 컴파일 오류 수정
