+++ src/Editor/EditorCore.h
@@ -18,6 +18,13 @@
 #include "Rendering/SkinnedMeshRegistry.h"
 #include "Editor/ViewportPicker.h"
 #include "Core/InputSystem.h"
+#include "Core/ReflectionUI.h"
+#include "Core/JsonRttr.h"
+#include "Core/ComponentRegistry.h"
+#include "imgui.h"
+#include <functional>
+#include <string>
+#include <memory>
 
 namespace Alice
 {
@@ -25,9 +32,60 @@ namespace Alice
 	class ResourceManager;
 	class SkinnedMeshRegistry;
 	class DeferredRenderSystem;
-	
-	// Undo/Redo 시스템 전방 선언 (EditorCore.cpp에서 정의됨)
-	struct ICommand;
+
+	// Undo/Redo 시스템
+	struct ICommand
+	{
+		virtual ~ICommand() = default;
+		virtual void Execute(World& world, EntityId& selectedEntity) = 0;
+		virtual void Undo(World& world, EntityId& selectedEntity) = 0;
+		virtual const char* GetDescription() const = 0;
+		virtual bool SupportsRedo() const { return true; }
+	};
+
+	// 컴포넌트 편집 명령 (템플릿)
+	template<typename T>
+	struct ComponentEditCommand : ICommand
+	{
+		EntityId entityId;
+		std::string componentTypeName;
+		JsonRttr::json oldJson;
+		JsonRttr::json newJson;
+		mutable std::string description;
+
+		ComponentEditCommand(EntityId id, const T& oldComp, const T& newComp)
+			: entityId(id), componentTypeName(rttr::type::get<T>().get_name().to_string())
+		{
+			rttr::instance oldInst = const_cast<T&>(oldComp);
+			rttr::instance newInst = const_cast<T&>(newComp);
+			oldJson = JsonRttr::ToJsonObject(oldInst);
+			newJson = JsonRttr::ToJsonObject(newInst);
+			description = "Edit " + componentTypeName;
+		}
+
+		void Execute(World& world, EntityId& selectedEntity) override
+		{
+			if (auto* comp = world.GetComponent<T>(entityId))
+			{
+				rttr::instance inst = *comp;
+				JsonRttr::FromJsonObject(inst, newJson);
+			}
+		}
+
+		void Undo(World& world, EntityId& selectedEntity) override
+		{
+			if (auto* comp = world.GetComponent<T>(entityId))
+			{
+				rttr::instance inst = *comp;
+				JsonRttr::FromJsonObject(inst, oldJson);
+			}
+		}
+
+		const char* GetDescription() const override
+		{
+			return description.c_str();
+		}
+	};
 
 	/// ImGui 컨텍스트 수명과 기본 에디터 유틸(도킹, 디렉터리 뷰, 에디터 패널 등)을 관리하는
 	/// 간단한 코어 클래스입니다.
@@ -71,7 +129,94 @@ namespace Alice
 
 		void DrawInspectorTransform(World& world, const EntityId& _selectedEntity);
 		void DrawInspectorScripts(World& world, const EntityId& _selectedEntity);
-		void DrawEngineComponent(const char* label, auto* comp, auto removeFn);
+		template<typename T>
+		void DrawEngineComponent(const char* label, T* comp, std::function<void()> removeFn, const EntityId& _selectedEntity, const std::string& compTypeName)
+		{
+			if (!comp) return;
+			if (ImGui::CollapsingHeader(label, ImGuiTreeNodeFlags_DefaultOpen)) {
+				std::string removeId = std::string("Remove##") + label;
+				if (ImGui::Button(removeId.c_str())) {
+					removeFn();
+					extern bool g_SceneDirty;
+					g_SceneDirty = true;
+					return;
+				}
+
+				// 컴포넌트 편집 이벤트 처리
+				static EntityId lastEditedEntity = InvalidEntityId;
+				static std::string lastEditedComponentType;
+				static JsonRttr::json editStartJson;
+
+				ReflectionUI::UIEditEvent event = ReflectionUI::RenderInspector(*comp);
+
+				// 편집 시작: oldJson 스냅샷 저장
+				if (event.activated && (_selectedEntity != lastEditedEntity || lastEditedComponentType != compTypeName))
+				{
+					rttr::instance inst = *comp;
+					editStartJson = JsonRttr::ToJsonObject(inst);
+					lastEditedEntity = _selectedEntity;
+					lastEditedComponentType = compTypeName;
+				}
+
+				// 편집 종료: newJson 저장하고 커맨드 푸시
+				if (event.deactivatedAfterEdit && _selectedEntity == lastEditedEntity && lastEditedComponentType == compTypeName)
+				{
+					rttr::instance inst = *comp;
+					JsonRttr::json editEndJson = JsonRttr::ToJsonObject(inst);
+
+					// 변경사항이 있으면 커맨드 푸시
+					if (editStartJson != editEndJson)
+					{
+						// 타입별로 적절한 커맨드 생성
+#define PUSH_COMPONENT_EDIT_CMD(T) \
+							if (compTypeName == rttr::type::get<T>().get_name().to_string()) \
+							{ \
+								T oldComp, newComp; \
+								rttr::instance oldInst = oldComp; \
+								rttr::instance newInst = newComp; \
+								JsonRttr::FromJsonObject(oldInst, editStartJson); \
+								JsonRttr::FromJsonObject(newInst, editEndJson); \
+								PushCommand(std::make_unique<ComponentEditCommand<T>>(_selectedEntity, oldComp, newComp)); \
+							}
+
+						// 주요 컴포넌트 타입들 처리
+						PUSH_COMPONENT_EDIT_CMD(MaterialComponent)
+							else PUSH_COMPONENT_EDIT_CMD(SkinnedMeshComponent)
+					else PUSH_COMPONENT_EDIT_CMD(SkinnedAnimationComponent)
+				else PUSH_COMPONENT_EDIT_CMD(CameraComponent)
+				else PUSH_COMPONENT_EDIT_CMD(CameraFollowComponent)
+				else PUSH_COMPONENT_EDIT_CMD(CameraSpringArmComponent)
+			else PUSH_COMPONENT_EDIT_CMD(CameraLookAtComponent)
+			else PUSH_COMPONENT_EDIT_CMD(CameraShakeComponent)
+						else PUSH_COMPONENT_EDIT_CMD(CameraBlendComponent)
+						else PUSH_COMPONENT_EDIT_CMD(CameraInputComponent)
+						else PUSH_COMPONENT_EDIT_CMD(PointLightComponent)
+						else PUSH_COMPONENT_EDIT_CMD(SpotLightComponent)
+						else PUSH_COMPONENT_EDIT_CMD(RectLightComponent)
+						else PUSH_COMPONENT_EDIT_CMD(ComputeEffectComponent)
+						else PUSH_COMPONENT_EDIT_CMD(Phy_RigidBodyComponent)
+						else PUSH_COMPONENT_EDIT_CMD(Phy_ColliderComponent)
+						else PUSH_COMPONENT_EDIT_CMD(Phy_MeshColliderComponent)
+						else PUSH_COMPONENT_EDIT_CMD(Phy_CCTComponent)
+						else PUSH_COMPONENT_EDIT_CMD(Phy_TerrainHeightFieldComponent)
+						else PUSH_COMPONENT_EDIT_CMD(Phy_JointComponent)
+
+#undef PUSH_COMPONENT_EDIT_CMD
+
+							extern bool g_SceneDirty;
+							g_SceneDirty = true;
+					}
+
+					lastEditedEntity = InvalidEntityId;
+					lastEditedComponentType.clear();
+				}
+
+				if (event.changed) {
+					extern bool g_SceneDirty;
+					g_SceneDirty = true;
+				}
+			}
+		}
 		void DrawInspectorMaterial(World& world, const EntityId& _selectedEntity);
 		void DrawInspectorPointLight(World& world, const EntityId& _selectedEntity);
 		void DrawInspectorSpotLight(World& world, const EntityId& _selectedEntity);
@@ -81,7 +226,7 @@ namespace Alice
 		// 물리잇
 		bool DrawLayerMaskEditor(const char* label, uint32_t& mask, const std::array<std::string, 32>& layerNames);
 		bool DrawLayerMaskChipEditor(const char* label, uint32_t& mask, const std::array<std::string, 32>& layerNames);
-		bool DrawIgnoreLayersChipEditor(const char* label, uint32_t& ignoreLayers, const std::array<std::string, 32>& layerNames);		
+		bool DrawIgnoreLayersChipEditor(const char* label, uint32_t& ignoreLayers, const std::array<std::string, 32>& layerNames);
 		void DrawInspectorCollider(World& world, const EntityId& _selectedEntity);
 		void DrawInspectorMeshCollider(World& world, const EntityId& _selectedEntity);
 		void DrawInspectorCharacterController(World& world, const EntityId& _selectedEntity);
@@ -106,7 +251,7 @@ namespace Alice
 		void EnsureSkinnedMeshesRegistered(World& world);
 		void SaveScene(World&);
 		void LoadScene(World&);
-		
+
 		// Undo 시스템
 		void PushCommand(std::unique_ptr<struct ICommand> cmd);
 
@@ -121,6 +266,3 @@ namespace Alice
 		bool               m_scriptBuilded = false;
 	};
 }
-
-
-
